Fix clippy suggestions

+++ src/binocle.rs
@@ -119,8 +119,8 @@ impl Binocle {
 
         for (i, pixel) in frame.chunks_exact_mut(4).enumerate() {
             let zoom_factor = settings.zoom_factor();
-            let x = (((i as isize) % WIDTH as isize) as isize) / zoom_factor;
-            let y = (((i as isize) / WIDTH as isize) as isize) / zoom_factor;
+            let x = ((i as isize) % WIDTH as isize) / zoom_factor;
+            let y = ((i as isize) / WIDTH as isize) / zoom_factor;
 
             let color = if x >= settings.width {
                 [0, 0, 0, 0]
